+++ source/common/upstream/cluster_manager_impl.cc
@@ -844,6 +844,11 @@ ThreadLocalCluster* ClusterManagerImpl::getThreadLocalCluster(absl::string_view 
   }
 }
 
+std::shared_ptr<FutureCluster>
+ClusterManagerImpl::futureThreadLocalCluster(absl::string_view cluster_name) {
+  return std::make_shared<ReadyFutureCluster>(cluster_name, *this);
+}
+
 void ClusterManagerImpl::maybePreconnect(
     ThreadLocalClusterManagerImpl::ClusterEntry& cluster_entry,
     const ClusterConnectivityState& state,
@@ -1539,6 +1544,31 @@ ProdClusterManagerFactory::createCds(const envoy::config::core::v3::ConfigSource
                                      ClusterManager& cm) {
   // TODO(htuch): Differentiate static vs. dynamic validation visitors.
   return CdsApiImpl::create(cds_config, cm, stats_, validation_context_.dynamicValidationVisitor());
+}
+
+ThreadLocalCluster* FutureCluster::getThreadLocalCluster() {
+  return cluster_manager_.getThreadLocalCluster(cluster_name_);
+}
+
+class DelayedFutureCluster::DumbHandle : public FutureCluster::Handle {
+public:
+  DumbHandle(DelayedFutureCluster& future) : future_(future) { ASSERT(!future_.is_canceled_); }
+  void cancel() { future_.is_canceled_ = true; }
+
+private:
+  DelayedFutureCluster& future_;
+};
+
+std::unique_ptr<FutureCluster::Handle> DelayedFutureCluster::await(Event::Dispatcher&,
+                                                                   ResumeCb cb) {
+  cb_ = [this, original_cb = std::move(cb)]() mutable {
+    if (!is_canceled_) {
+      original_cb(*this);
+    }
+    // Erase the function to release the captures.
+    original_cb = nullptr;
+  };
+  return std::make_unique<DumbHandle>(*this);
 }
 
 } // namespace Upstream
