+++ source/common/upstream/cluster_manager_impl.h
@@ -25,6 +25,7 @@
 #include "envoy/thread_local/thread_local.h"
 #include "envoy/upstream/cluster_manager.h"
 
+#include "common/common/assert.h"
 #include "common/common/cleanup.h"
 #include "common/config/grpc_mux_impl.h"
 #include "common/config/subscription_factory_impl.h"
@@ -253,6 +254,8 @@ public:
 
   const ClusterSet& primaryClusters() override { return primary_clusters_; }
   ThreadLocalCluster* getThreadLocalCluster(absl::string_view cluster) override;
+
+  std::shared_ptr<FutureCluster> futureThreadLocalCluster(absl::string_view cluster_name) override;
 
   bool removeCluster(const std::string& cluster) override;
   void shutdown() override {
@@ -604,6 +607,47 @@ private:
 
   Config::SubscriptionFactoryImpl subscription_factory_;
   ClusterSet primary_clusters_;
+};
+
+class ReadyFutureCluster : public FutureCluster {
+public:
+  ReadyFutureCluster(absl::string_view cluster_name, ClusterManager& cluster_manager)
+      : FutureCluster(cluster_name, cluster_manager) {}
+  bool isReady() override { return true; }
+  std::unique_ptr<Handle> await(Event::Dispatcher&, ResumeCb) override {
+    // User should not call await since ReadyFutureCluster is always ready.
+    NOT_IMPLEMENTED_GCOVR_EXCL_LINE;
+  }
+};
+
+// This future cluster readiness is controlled by the not-owning flag. It can be used by emulate
+// never ready future cluster or eventually ready cluster.
+class DelayedFutureCluster : public FutureCluster {
+public:
+  DelayedFutureCluster(absl::string_view cluster_name, ClusterManager& cluster_manager,
+                       bool& ready_flag)
+      : FutureCluster(cluster_name, cluster_manager), ready_(ready_flag) {}
+  // FutureCluster
+  bool isReady() override { return ready_; }
+  std::unique_ptr<Handle> await(Event::Dispatcher&, ResumeCb cb) override;
+
+  // Invoke the callback.
+  void readyCallback() {
+    ASSERT(ready_);
+    cb_();
+  }
+
+private:
+  class DumbHandle;
+  friend class DumbHandle;
+  // The callback which is optional invoked on the target dispatcher.
+  std::function<void()> cb_;
+  // The flag if the future cluster is ready. This flag should only be switched from false to true.
+  bool& ready_;
+  // The flag if the ResumeCb is valid. This flag should only be switch false to true by handler.
+  // There is no mutex protecting this flag so the mutation must be executed on the same thread
+  // where the callback is invoked.
+  bool is_canceled_{false};
 };
 
 } // namespace Upstream
